# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BreakBPM API — accounts, passes, and game persistence
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscriptions/checkout:
    post:
      operationId: createSubscriptionCheckout
      tags: [passes]
      summary: Begin a subscription purchase — returns a checkout URL + token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubscriptionCheckoutInput" }
      responses:
        "200":
          description: Checkout created (or rejected if no provider / Lifetime held)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckoutResult" }
  /subscriptions/verify:
    post:
      operationId: verifySubscriptionCheckout
      tags: [passes]
      summary: Verify a subscription checkout token and activate on success
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerifyCheckoutInput" }
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubscriptionVerifyResult" }
...
  /subscriptions/dev-activate:
    post:
      operationId: devActivateSubscription
      tags: [passes]
      summary: (DEV ONLY) Activate a free subscription for the signed-in user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DevActivateSubscriptionInput" }
      responses:
        "200":
          description: Activated (or rejected)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubscriptionVerifyResult" }
...
components:
  schemas:
...
    SubscriptionSummary:
      type: object
      required: [status, interval, currentPeriodEnd, cancelAtPeriodEnd]
      properties:
        status: { type: string, enum: [active, past_due, canceled] }
        interval: { type: string, enum: [month, year] }
        # Paid-through date. Renews on this date unless cancelAtPeriodEnd is
        # set, in which case access ends here.
        currentPeriodEnd: { type: string, format: date-time }
        cancelAtPeriodEnd: { type: boolean }
...
    SubscriptionCheckoutInput:
      type: object
      required: [interval]
      properties:
        interval: { type: string, enum: [month, year] }
    SubscriptionVerifyResult:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean }
        message: { type: string }
        subscription: { $ref: "#/components/schemas/SubscriptionSummary" }
...
    DevActivateSubscriptionInput:
      type: object
      required: [interval]
      properties:
        interval: { type: string, enum: [month, year] }
...
    CheckoutResult:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean }
        message: { type: string }
        # Provider-issued opaque token. Hand this to /passes/verify to
        # complete the purchase. Absent when success=false.
        opaqueToken: { type: string }
        # Optional URL to redirect the user to (e.g. Stripe Checkout).
        checkoutUrl: { type: string }
    VerifyCheckoutInput:
      type: object
      required: [opaqueToken]
      properties:
        opaqueToken: { type: string, minLength: 1, maxLength: 512 }
...