# lib/api-zod/src/generated/types/devActivateSubscriptionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BreakBPM API — accounts, passes, and game persistence
 * OpenAPI spec version: 0.1.0
 */
import type { DevActivateSubscriptionInputInterval } from './devActivateSubscriptionInputInterval';

export interface DevActivateSubscriptionInput {
  interval: DevActivateSubscriptionInputInterval;
}