# lib/api-zod/src/generated/types/subscriptionVerifyResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BreakBPM API — accounts, passes, and game persistence
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriptionSummary } from './subscriptionSummary';

export interface SubscriptionVerifyResult {
  success: boolean;
  message: string;
  subscription?: SubscriptionSummary;
}